refactor: split god files + fix backend stubs and security
Backend:
- Fix hardcoded Guardian JWT secret — now reads SECRET_KEY_BASE from env
- Fill computed values in agent, cost, dashboard, schedule, signal controllers
- Mark stub adapters (cursor, gemini, openclaw) with real behaviour skeleton
- Wire SpawnController to actually create agent processes

Desktop:
- Split onboarding/+page.svelte (1,553→7 step components)
- Split settings/+page.svelte (1,057→5 tab components)
- Split HireAgentDialog (628→4 sub-components)
- Split mock/index.ts (669→17 domain-specific mock files)

All codebases compile clean (0 errors)
